+++ EssentialFeed/EssentialFeediOS/Composer/FeedUIComposer.swift
@@ -37,27 +37,6 @@ private extension FeedViewController {
     }
 }
 
-final class WeakRefVirtualProxy<T: AnyObject> {
-    weak var object: T?
-    
-    init(_ object: T) {
-        self.object = object
-    }
-}
-
-extension WeakRefVirtualProxy: FeedLoadingView where T: FeedLoadingView {
-    func display(_ viewModle: FeedLoadingViewModel) {
-        object?.display(viewModle)
-    }
-}
-
-extension WeakRefVirtualProxy: FeedImageView where T: FeedImageView, T.Image == UIImage {
-    func display(_ viewModel: FeedImageViewModel<UIImage>) {
-        object?.display(viewModel)
-    }
-}
-
-
 final class FeedViewAdapter: FeedView {
     weak var controller: FeedViewController?
     let loader: FeedImageDataLoader
